move notify code back into proxy.go
