--- server/proxy/notify.go
@@ -1,192 +1,0 @@
-package proxy
-
-import (
-	"encoding/json"
-	"path"
-
-	"github.com/mattermost/mattermost-plugin-apps/apps"
-	"github.com/mattermost/mattermost-plugin-apps/upstream"
-	"github.com/mattermost/mattermost-plugin-apps/utils"
-
-	"github.com/mattermost/mattermost-server/v5/model"
-	"github.com/pkg/errors"
-)
-
-func (p *Proxy) Notify(cc *apps.Context, subj apps.Subject) error {
-	subs, err := p.store.Subscription.Get(subj, cc.TeamID, cc.ChannelID)
-	if err != nil {
-		return err
-	}
-
-	return p.notify(cc, subs)
-}
-
-func (p *Proxy) NotifyRemoteWebhook(app *apps.App, data []byte, webhookPath string) error {
-	if !app.GrantedPermissions.Contains(apps.PermissionRemoteWebhooks) {
-		return utils.NewForbiddenError("%s does not have permission %s", app.AppID, apps.PermissionRemoteWebhooks)
-	}
-
-	up, err := p.upstreamForApp(app)
-	if err != nil {
-		return err
-	}
-
-	var datav interface{}
-	err = json.Unmarshal(data, &datav)
-	if err != nil {
-		// if the data can not be decoded as JSON, send it "as is", as a string.
-		datav = string(data)
-	}
-
-	// TODO: do we need to customize the Expand & State for the webhook Call?
-	creq := &apps.CallRequest{
-		Call: apps.Call{
-			Path: path.Join(apps.PathWebhook, webhookPath),
-		},
-		Context: p.conf.GetConfig().SetContextDefaultsForApp(app.AppID, &apps.Context{
-			ActingUserID: app.BotUserID,
-		}),
-		Values: map[string]interface{}{
-			"data": datav,
-		},
-	}
-	expander := p.newExpander(creq.Context, p.mm, p.conf, p.store, "")
-	creq.Context, err = expander.ExpandForApp(app, creq.Expand)
-	if err != nil {
-		return err
-	}
-
-	return upstream.Notify(up, creq)
-}
-
-func (p *Proxy) NotifyMessageHasBeenPosted(post *model.Post, cc *apps.Context) error {
-	postSubs, err := p.store.Subscription.Get(apps.SubjectPostCreated, cc.TeamID, cc.ChannelID)
-	if err != nil && err != utils.ErrNotFound {
-		return errors.Wrap(err, "failed to get post_created subscriptions")
-	}
-
-	subs := postSubs
-	mentions := model.PossibleAtMentions(post.Message)
-
-	botCanRead := map[string]bool{}
-	if len(mentions) > 0 {
-		appsMap := p.store.App.AsMap()
-		mentionSubs, err := p.store.Subscription.Get(apps.SubjectBotMentioned, cc.TeamID, cc.ChannelID)
-		if err != nil && err != utils.ErrNotFound {
-			return errors.Wrap(err, "failed to get bot_mentioned subscriptions")
-		}
-
-		for _, sub := range mentionSubs {
-			app := appsMap[sub.AppID]
-			if app == nil {
-				continue
-			}
-			for _, mention := range mentions {
-				if mention == app.BotUsername {
-					_, ok := botCanRead[app.BotUserID]
-					if ok {
-						// already processed this bot for this post
-						continue
-					}
-
-					canRead := p.canReadChannel(app.BotUserID, post.ChannelId)
-					botCanRead[app.BotUserID] = canRead
-
-					if canRead {
-						subs = append(subs, sub)
-					}
-				}
-			}
-		}
-	}
-
-	if len(subs) == 0 {
-		return nil
-	}
-
-	return p.notify(cc, subs)
-}
-
-func (p *Proxy) NotifyUserHasJoinedChannel(cc *apps.Context) error {
-	return p.notifyJoinLeave(cc, apps.SubjectUserJoinedChannel, apps.SubjectBotJoinedChannel)
-}
-
-func (p *Proxy) NotifyUserHasLeftChannel(cc *apps.Context) error {
-	return p.notifyJoinLeave(cc, apps.SubjectUserLeftChannel, apps.SubjectBotLeftChannel)
-}
-
-func (p *Proxy) NotifyUserHasJoinedTeam(cc *apps.Context) error {
-	return p.notifyJoinLeave(cc, apps.SubjectUserJoinedTeam, apps.SubjectBotJoinedTeam)
-}
-
-func (p *Proxy) NotifyUserHasLeftTeam(cc *apps.Context) error {
-	return p.notifyJoinLeave(cc, apps.SubjectUserLeftTeam, apps.SubjectBotLeftTeam)
-}
-
-func (p *Proxy) notifyJoinLeave(cc *apps.Context, subject, botSubject apps.Subject) error {
-	userSubs, err := p.store.Subscription.Get(subject, cc.TeamID, cc.ChannelID)
-	if err != nil && err != utils.ErrNotFound {
-		return errors.Wrapf(err, "failed to get %s subscriptions", subject)
-	}
-
-	botSubs, err := p.store.Subscription.Get(botSubject, cc.TeamID, cc.ChannelID)
-	if err != nil && err != utils.ErrNotFound {
-		return errors.Wrapf(err, "failed to get %s subscriptions", botSubject)
-	}
-
-	subs := userSubs
-	appsMap := p.store.App.AsMap()
-	for _, sub := range botSubs {
-		app := appsMap[sub.AppID]
-		if app == nil {
-			continue
-		}
-
-		if app.BotUserID == cc.UserID {
-			subs = append(subs, sub)
-		}
-	}
-
-	return p.notify(cc, subs)
-}
-
-func (p *Proxy) notify(cc *apps.Context, subs []*apps.Subscription) error {
-	expander := p.newExpander(cc, p.mm, p.conf, p.store, "")
-
-	notify := func(sub *apps.Subscription) error {
-		call := sub.Call
-		if call == nil {
-			return errors.New("nothing to call")
-		}
-
-		callRequest := &apps.CallRequest{Call: *call}
-		app, err := p.store.App.Get(sub.AppID)
-		if err != nil {
-			return err
-		}
-		callRequest.Context, err = expander.ExpandForApp(app, callRequest.Expand)
-		if err != nil {
-			return err
-		}
-		callRequest.Context.Subject = sub.Subject
-
-		up, err := p.upstreamForApp(app)
-		if err != nil {
-			return err
-		}
-		return upstream.Notify(up, callRequest)
-	}
-
-	for _, sub := range subs {
-		err := notify(sub)
-		if err != nil {
-			p.mm.Log.Debug("Error sending subscription notification to app", "app_id", sub.AppID, "subject", sub.Subject, "err", err.Error())
-		}
-	}
-
-	return nil
-}
-
-func (p *Proxy) canReadChannel(userID, channelID string) bool {
-	return p.mm.User.HasPermissionToChannel(userID, channelID, model.PERMISSION_READ_CHANNEL)
-}
